Fixed Normalized images bug

+++ utils.py
@@ -38,11 +38,12 @@ def save_predictions_as_imgs(loader, model, folder='imgs/', device='cuda'):
         with torch.no_grad():
             preds = torch.sigmoid(model(x))
             preds = (preds > 0.5).float()
-            
+
         torchvision.utils.save_image(
             preds, f'{folder}/pred_{idx}.png'
-        )
+        )       
+
         torchvision.utils.save_image(
-            y.unsqueeze(1), f'{folder}/mask_{idx}.png'
+            y.float().unsqueeze(1), f'{folder}/mask_{idx}.png'
         )
         #print(y.shape)
